Add connection resiliency support for Sql Server. All high-level API that need database connection will use the configured execution strategy except migrations. Handle external connection state changes in RelationalConnection
Handle more objects in SqlServerDatabaseCleaner
Functional tests are no longer executed in a fixture-wide transaction, the database is cleaned when a fixture is created. Tests that modify a shared database do so in a local transaction
All SQL Server functional tests can be run against SQL Azure by providing the connection string in config.json, database will be created using Basic service tier or in an elastic pool if specified.

Add to .Core:
  IExecutionStrategy
  ExecutionStrategy
  IExecutionStrategyFactory
  IDatabaseProviderServices
    IExecutionStrategyFactory ExecutionStrategyFactory
  DatabaseFacade
    IExecutionStrategy CreateExecutionStrategy()
    TResult ExecuteWithStrategyInTransaction<TResult>(Func<DbContext, TResult> operation)
    Task<TResult> ExecuteWithStrategyInTransactionAsync<TResult>(Func<DbContext, CancellationToken, Task<TResult>> operation, CancellationToken cancellationToken)

Add to .Relational:
  RelationalDbContextOptionsBuilder
    ExecutionStrategy(Func<ExecutionStrategyContext, IExecutionStrategy> getExecutionStrategy)

Add to .SqlServer:
  SqlAzureExecutionStrategy
  SqlServerDbContextOptionsBuilder
    EnableRetryOnFailure()
    EnableRetryOnFailure(int maxRetryCount)

asgasg